requirements.txt
Flask
python-dateutil
flup
beautifulsoup4
argparse